# docs/06-api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders/{id}/accept": {
      "post": {
        "operationId": "OrdersController_accept",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order, now MERCHANT_ACCEPTED"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not a merchant"
          },
          "409": {
            "description": "INVALID_TRANSITION or NOT_RESTAURANT_MEMBER"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders/{id}/pickup": {
      "post": {
        "operationId": "OrdersController_pickup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order, now PICKED_UP"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not a rider"
          },
          "409": {
            "description": "INVALID_TRANSITION"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/orders/{id}/complete": {
      "post": {
        "operationId": "OrdersController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order, now DELIVERED"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not a rider"
          },
          "409": {
            "description": "INVALID_TRANSITION"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "OrdersController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The order, now CANCELLED"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the order owner and not an operator"
          },
          "404": {
            "description": "Order not found, or not owned by this customer"
          },
          "409": {
            "description": "INVALID_TRANSITION — past the cancellable window"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "orders"
        ]
      }
    },
...
    "/api/v1/rider/location": {
      "post": {
        "operationId": "RiderController_updateLocation",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The position was recorded"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not an approved rider"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rider"
        ]
      }
    },
    "/api/v1/rider/offers/{id}/accept": {
      "post": {
        "operationId": "RiderController_acceptOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The delivery, now RIDER_ASSIGNED to this rider"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not an approved rider"
          },
          "404": {
            "description": "Offer not found, or not offered to this rider"
          },
          "409": {
            "description": "OFFER_TAKEN, OFFER_EXPIRED, or RIDER_HAS_ACTIVE_DELIVERY"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rider"
        ]
      }
    },
    "/api/v1/rider/offers/{id}/decline": {
      "post": {
        "operationId": "RiderController_declineOffer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The offer, now DECLINED"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not an approved rider"
          },
          "404": {
            "description": "Offer not found, or not offered to this rider"
          },
          "409": {
            "description": "OFFER_TAKEN or OFFER_EXPIRED — the offer is no longer PENDING"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "rider"
        ]
      }
    },
...
  },
  "info": {
    "title": "BANHAO API",
    "description": "BANHAO | บ้านเฮา — Phase 1 Food Delivery API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}